Rendering of fluids in pipes completed.

+++ src/minecraft/ip/industrialProcessing/transport/fluids/IPressuredTank.java
@@ -3,5 +3,5 @@ package ip.industrialProcessing.transport.fluids;
 import net.minecraftforge.common.ForgeDirection;
 
 public interface IPressuredTank {
-	int getPressure(ForgeDirection from);
+	float getPressure(ForgeDirection from);
 }
